Move cache to new file
