+++ aggregator/src/lib.rs
@@ -3,6 +3,7 @@
 
 pub mod aggregator;
 pub mod binary_utils;
+pub mod cache;
 pub mod config;
 pub mod metrics;
 pub mod trace;
